+++ src/index.mjs
@@ -2,6 +2,7 @@
 
 import { createState } from './state/state.mjs'
 import { html } from './template/template.mjs'
+import { customElement } from './template/custom_element.mjs'
 import Router from './router/router.mjs'
 import Component from './component/component.mjs'
 import I18n from './intl/i18n.mjs'
@@ -11,5 +12,6 @@ export {
   Router,
   Component,
   I18n,
-  html
+  html,
+  customElement
 }