+++ test/index.mjs
@@ -1,3 +1,4 @@
+import './template/custom_element.mjs'
 import './template/template.mjs'
 import './router/router.mjs'
 import './component/component.mjs'